fix(staging-mode): Make rollback semantics explicit for all ContainerMessageTypes in light of Staging Mode (#24727)
Only one behavior change here: Block setting aliases while in Staging
Mode. This is a consensus operation and not well-suited for Staging
Mode. We don't expect any partner will have a need for this, but we can
revisit if/when they do.

Besides that, just making things a little more explicit and adding some
tests.