Add support for kafka consumer cooperative rebalancing for node-rdkafka client (#24743)
Add support for kafka consumer cooperative rebalancing for node-rdkafka
client.

## Description

With the default rebelancing approach, kafka consumer group coordinator
would disconnect all the consumers when rebalancing is needed by sending
a revoke message with the list of all currently assigned partitions.
This action is also known as "stop-the-world" would effectively pause a
consumer affecting latency and throughput. Then, the coordinator will
send a list of newly assigned partitions per consumer. Having more than
a few consumers in a group amplifies the problem as any join/leave
action triggers "stop-the-world" state for the group.

The cooperative rebalancing doesn't require "stop-the-world" state as it
would revoke only those partitions that will be indeed moved to another
consumer, while the partitions that would have been returned back to it
in case of "stop-the-world" continue to operate without disruption.