refactor(api-markdown-documenter): Update Documentation Domain to be more restrictive (#24667)
### Background

The Documentation Domain (`DocumentationNode` and its implementations)
was created with a goal of being very flexible in the kinds of trees of
content it could express. This was accomplished by, in most cases,
allowing parent nodes to contain any kind of child content (with a
couple of exceptions). This ultimately ended up creating problems in
terms of Markdown-compatibility in particular (HTML is more expressive,
so compat there was less of an issue). There are also some forms of
hierarchy allowed by the system (and inadvertently used by the system)
that don't make sense (e.g., parenting a `SectionNode` under a
`ParagraphNode`.

As this library has evolved, I have become more and more convinced that
the Documentation Domain is unnecessary and should be removed. In this
world, `ApiItem` transformations would output `mdast` (Markdown Syntax
Trees) directly, instead of going through an intermediary domain.
`mdast` has a vast support ecosystem, so it seems like a much better
target than some bespoke intermediary domain.

### This PR

This PR is an attempt to begin aligning the Documentation Domain with
Markdown syntax requirements by restricting the kinds of contents that
can appear under specific kinds of nodes.

It also introduces a new pattern for domain extension, which itself
aligns with how the broader `unist` ecosystem accomplishes
extensibility. We require this extensibility internally, as our
`fluidframework.com` website build actually does extend the
Documentation Domain with a custom node type. In the future, this can be
trivially mapped to a custom `mdast` node, but for now we need to
support it.

#### Immediate impact

See the updated HTML test assets for the immediate impact on the output
format. Note the reduction in unnecessary hierarchy.