refactor(api-markdown-documenter): Make `SpanNode`s more restrictive (#24750)
We use `SpanNode` to represent a block of content with common formatting
applies. It can be thought of as sort of an amalgamation of italics,
bold, and underline formatting components (which in other contexts like
HTML would be separate node kinds).

Previously, this type supported wrapping block content (paragraphs,
etc.) in common formatting. This is a departure from Markdown / MDAST.
It turns out we only used this support in a single place, and that was
for strictly subjective formatting reasons (which are ultimately unused
anyways).

This PR brings the span concept in line with Markdown requirements by
restricting SpanNode children to phrasing content only.