fix(staging-mode): Add some more preconditions to enterStagingMode (#24738)
Adding these preconditions to `enterStagingMode`:

1. Must not be detached.
- Changes made while detached are local-only already, and typically
don't result in ops anyway (included in attach summary instead), so this
is not a useful mode. So codify this assumption to simplify our own
reasoning about Staging Mode
2. Must be using TurnBased flush mode
- TurnBased is the only option we provide, with one caveat which is 2.x
Declarative Mode clients that want to support collab with 1.x clients.
We don't anticipate offering Staging Mode in that configuration.