fix(container-runtime): Rename false `connected` checks in runtime to `canSendOps` (#24394)
## Description

In setConnectionState(), the current `connected` sent from the container
to the container runtime is not actually the connection state of the
container, but rather a boolean that tells us whether the container can
send ops (connected && !readonly).

This change renames `connected` usage in ContainerRuntime to
`canSendOps` to clarify the true meaning of the variable.