improvement(staging-mode): Clarify Squash/Rollback behavior for ContainerMessageType.GC (#24774)
GC ops can be made while in Staging Mode, and we need to be very clear
with how they're handled.

* **Squashing:** They don't need to be squashed, no personal data to
protect
* **Rollback:** GC ops submitted while in Staging Mode will be dropped
if staged changes are discarded. (these are all "Tombstone Loaded" ops,
which prevent data loss in case of invalid GC data)
* We will log a new error event `GC_OpDiscarded` to highlight this case
* We are also adding a prop `inStagingMode` to all ContainerRuntime
logs, including `GC_Tombstone_DataStore_Requested` etc, to further aid
investigations

Also refactor the Resubmit code so it's clear that being able to support
squash is an exceptional case - most op types don't need to deal with
it.