fix(Staging-Mode): Handle rollback case for DocumentSchemaMessages (#24776)
A DocumentSchema message could be sent in Staging Mode, which means we
need to handle what happens if it's rolled back.

Turns out it's easy - no internal state changes when _generating_ the op
to be submitted, only when _processing_. Except for a boolean flag
indicating whether we are waiting for a pending op, or if we should
still try to generate one. So all that's needed is to reset this flag
when rolling back.