improvement(api-markdown-documenter): Simplify Markdown table rendering (#24782)
Updates Markdown rendering to use only a single `-` for heading
delimiter cells, rather than 3. This reduces the size of generated
documents, but more importantly, it aligns with
`mdast-util-to-markdown`'s behavior. The plan is to replace our custom
Markdown rendering logic with this library, so making this change now
will make future PR diffs easier to reason about and review.