Interval Collections: Deprecate unnecessary and internal APIs (#24792)
Deprecate unnecessary and internal APIs in `ISequenceIntervalCollection`
and related interval types.

The following APIs are now deprecated and will be removed in a future
release:
- `IInterval.clone`
- `IInterval.modify`
- `IInterval.union`
- `ISerializableInterval`
- `SequenceInterval.clone`
- `SequenceInterval.modify`
- `SequenceInterval.union`
- `SequenceInterval.serialize`
- `SequenceInterval.addPositionChangeListeners`
- `SequenceInterval.removePositionChangeListeners`


These APIs were never intended for public use. There is no migration
path, and any usage is strongly discouraged, as it may result in severe
errors or data corruption. Please remove any dependencies on these APIs
as soon as possible.

---------

Co-authored-by: jzaffiro <110866475+jzaffiro@users.noreply.github.com>