Interval Collection: Consolidate local seq handling (#24811)
This change consolidates the local seq handling to submit and process
and combines localSeqToRebasedInterval with
localSeqToSerializedInterval. The primary motivation here is refactoring
to simplify local state handling as part of a larger refactoring.
however, this change also fixes a memory leak related to
localSeqToRebasedInterval, as the previous code never cleaned up local
seqs in that map.