Remove LazyField Iterator (#24817)
## Description

This iterator was not exposed on the interface, and only had one use
which could be better optimized by inlining the implementation.

This change was motivated by wanting to clear up room for adding a
different iterator to the FlexTreeField interface to help it align with
MapTree like how UnhydratedFlexTreeField does.