improvement(api-markdown-documenter): Don't needlessly escape `>` characters in Markdown (#24813)
Escaping of `>` is not strictly required in Markdown. `>` only has
special meaning if it is closing an unescaped `<`. So, as long as we
continue to escape the latter, we don't need to escape the former.

The primary goal of this PR is to align our Markdown rendering with the
behavior of `mdast-util-to-markdown`, which we plan to use to replace
our custom Markdown rendering.