(tree) Renamed some shape encoding entities in chunked forest (#24777)
Renamed the following related to shape encoding in chunked forest:
- `handleShapesAndIdentifiers` -> `updateShapesAndIdentifiersEncoding` -
This makes it explicit that the function update the encoding of shapes
and identifiers in the passed buffer.
- `count` in `Shape` -> `countReferencedShapesAndIdentifiers` - This
makes it clear that this function counts any shapes and identifiers in
the `Shape` and counts them.
- `shapes` paramater in the above function -> `shapeDiscovered` - This
makes it explicit that this is a callback which should be called
whenever a shape is discovered.