Interval Collection: Rename IMapMessageLocalMetadata  (#24823)
Long ago interval collections were based on map, and when they were
split, a number of types were copied. This change renames
IMapMessageLocalMetadata to IntervalMessageLocalMetadata so it better
represents what is used for. Additionally, I've added a type to
IntervalMessageLocalMetadata such that they can be discriminated, and
each can have its own metadata, which reduces our reliance on optional
properties and makes the code clearer.