refactor(aqueduct): Hide `sharedTree` property on `TreeDataObject` (#24835)
This property was exposed ultimately for access by Devtools. If a user
wants direct access to the `SharedTree` object, they can make that their
(optionally) derived view, so exposing this publicly isn't really needed
outside of devtools. Instead, devtools now takes the same approach for
accessing this property as it does for `DataObject`.