(tree) Unify summarize and getAttachSummary methods on summarizables (#24761)
## Description
This PR makes the following changes:
1. Removed the `getAttachSummary` method from `Summarizable` in shared
tree and replaces it with the existing `summarize` method.
These methods are supposed to be used in two different scenarios -
`getAttachSummary` for the summary that is synchronously generated
during container attach and `summarize` for the summary that is
asynchronously generated during regular summarization in case tree wants
to support async summarization for scenarios like virtualization.
However, currently only `getAttachSummary` is used in both these
scenarios since DDSes do not do async summarization. Some of the
summarizables implement both of these methods and some only implement
one of them. This leads to some confusion and also using the method
"getAttachSummary" for regular summarization is also misleading.
So, this PR keeps the `summarize` method and makes in syncronous. The
async version of this method can be added later when it is needed.
2. Added a `fullTree` param to the `summarizeCore` method in
`SharedObject`. This is needed for tree (or any DDS) to support
incremental summarization. If this param is true, a full tree must be
generated without any incrementality.
3. In `SharedTreeCore`'s `summarizeCore` method, append the sub-tree
path's to the incremental summary context before passing it to the
individual summarizables. This will keep the logic in one place without
having each summarizable do it and also, the summarizables shouldn't
know about its parent's incremental summary details.


[AB#40395](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/40395)