Update CSP header and hash external scripts (#24828)
Seeing some eval violations in the CREM board. Following the same
reasoning of cleaning up violations, even though we are not required to
meet all of them, I'm adding a couple of options to ignore those
violations. According to
https://www.owiki.ms/wiki/Strict_Content_Security_Policy, 'unsafe-eval'
is allowed while 'unsafe-inline' would be ignored in the presence of
hashes, which we have. Adding all options referenced in the wiki anyway
to be as close as possible as recommended header example. The only
exception would be 'strict-dynamic' which disables host-based
allowlisting