fix(changelog-generator-wrapper): Fix imports and metadata (#24833)
While using the changelog tool from the server release group, I noticed
that pnpm was only linking index.js, so the imports weren't there when
the code ran, causing an import error.

As best as I can tell, pnpm analyzes the files that will be produced by
the package and links only those, so by removing the `files` key in
package.json, the whole contents of the package should be linked. The
package is not published so there is little concern.

I also updated the imports to have a file extension, since that is the
modern node style.