(compat) Fixed layer compat test failing due to UsageError validation (#24858)
## Bug
Some of the layer compat end-to-end tests were disabled recently via
https://github.com/microsoft/FluidFramework/pull/24843.

These tests were failing in the end-to-end test CI pipeline. The reason
for the failures was that they were validating that the error thrown due
to layer compat incompatibility were an instance of UsageError. However,
there are 2 classes of UsageError loaded (it has 2 implementations) and
so, for some reason due to either compat config or some other dynamic
import, there was a mismatch during validation.

## Fix
This PR fixes the issue by validating the error is `isFluidError` and
the errorType is `usageError` asn these checks are agnostic to
instanceOf.