improvement(test-end-to-end-tests): Use minVersionForCollab in e2e tests (#24810)
## Description

This PR updates e2e tests to prepare for the removal of
`filterRuntimeOptionsForVersion()` in
https://github.com/microsoft/FluidFramework/pull/24809. This PR is a
subset of https://github.com/microsoft/FluidFramework/pull/24809 and
only contains the changes to tests.

The tests updated in this PR are updated for one of these two reasons:

1. Some runtime options that were automatically set by
`filterRuntimeOptionsForVersion()` are not automatically set by
`minVersionForCollab` . For example, `enableRuntimeIdCompressor` will
never be enabled by default by any `minVersionForCollab`, but was
enabled automatically by `filterRuntimeOptionsForVersion()`. In these
cases, we need to manually enable `enableRuntimeIdCompressor` (if not
already done by the test author).

2. Second, some tests that are enabled for `"FullCompat"` manually
enable runtime options that collide the automatically assigned
`minVersionForCollab`. For example, some tests in `compression.spec.ts`
have batching/compression enabled. When these tests are running in
cross-client compat mode, `minVersionForCollab` will be set as low as
`1.4.0` (for the new client). This results in an error thrown unless the
test explicitly sets the `minVersionForCollab` to a higher value.
For more context on why we throw errors in this scenario, see
https://github.com/microsoft/FluidFramework/pull/24625.

## Misc


[AB#41365](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/41365)