Tag command injection fix (#24861)
## Description

Use checkout action instead of command line for checking out a tag in
GitHub Actions.

I validated the bug fix by putting a code injection in release tag and
then pushing the tag to main on my fork. The result of this was that the
injection was completely omitted from the tag name, and the workflow
worked as expected:

![image](https://github.com/user-attachments/assets/0507d8ee-fee4-4f10-a882-8ffeb0a40af9)

https://github.com/Toritos01/FluidFramework/actions/runs/15691947672/job/44209020812

---------

Co-authored-by: Brandon Diaz <“BrandonLouisDiaz@gmail.com”>