refactor(presence): Refactor API to prepare for runtime data validators (#24432)
Updates the presence API to prepare for runtime validation.

## Summary

- The `StateFactory.latest` and `StateFactory.latestMap` functions now
accept a new argument, `validator`, which can be used to pass a runtime
validator function to the presence system.
- The same functions now return a `Latest`/`LatestMap` when a validator
is provided; otherwise they return a `LatestRaw`/`LatestMapRaw`. This is
implemented in a single class with function overloads to narrow the
types based on the presence of the `validator` argument.
- Utility types were added to add variations to the `*Events` types so a
single event type is used for both raw and non-raw data types. The
RawEvents types remain so customers do need to work with system
`ValueAccessor` types.
- These same utility types are used in other interfaces to vary their
behavior so the type system represents the resulting data correctly
based on usage.
- Some basic test code was added to verify that the types returned by
the function vary based on the presence or absence of the `validator`
argument.

---------

Co-authored-by: Jason Hartman <jasonha@microsoft.com>