improvement(test-version-utils): Automatically set minVersionForCollab in e2e test generation (#24809)
## Description

This PR leverages `minVersionForCollab` in e2e test generations by doing
the following:
1. Allows test authors to set `minVersionForCollab` from the
`ITestContainerConfig` object.
2. If no `minVersionForCollab` is provided in `ITestContainerConfig`,
then it will be set automatically. We will use the lesser of the
container runtime version for creating/loading. (Note: If it's not a
cross-client test, then the loading version will be undefined and we
will use the creating version).

These changes allow us to remove the `filterRuntimeOptionsForVersion()`
function. `filterRuntimeOptionsForVersion` was used to set runtime
options based on the container runtime verison, which is no longer
necessary since that logic is automatically taken care of by
`minVersionForCollab`.

## Misc


[AB#41365](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/41365)