docs(presence): Centralize docs and update changeset (#24882)
The presence docs page and the readme are now connected via an md-magic
include. This is a brittle form of reuse, but it should at least raise
alarms and whatnot if the content starts to diverge. I set up the
include on the README. That is, the presence.mdx file is the source of
truth. Docs updates there should be reproduced in the README by running
`pnpm run build:readme` from the repo root.

I also updated the docs to make some corrections and wording fixes, and
updated the changeset for the upcoming release.

---------

Co-authored-by: jzaffiro <110866475+jzaffiro@users.noreply.github.com>