Limit permissions made available to tokens in GitHub actions (#24920)
## Description

These changes were made to limit the permissions available to tokens in
GitHub actions, so that we can improve our score for Token-Permissions
on scorecard.

The changes made were done by following guidelines from the
Token-Permissions section of the scorecard docs:

https://github.com/ossf/scorecard/blob/ea7e27ed41b76ab879c862fa0ca4cc9c61764ee4/docs/checks.md#token-permissions
Mainly, write permissions were moved to the job level, with read
permissions at the top-level.

This tool (recommended by the scorecard docs) was used to help determine
the minimum permissions needed on the workflow:
https://app.stepsecurity.io/secureworkflow

---------

Co-authored-by: Brandon Diaz <“BrandonLouisDiaz@gmail.com”>