Add missing top-level permissions to GitHub Actions (#24928)
## Description

Add top-level permissions to GitHub Actions workflows that already have
job level permissions, so that we can maximize our scorecard rating.
They deduct one point if a top-level permission is not defined, even if
it is not needed; this is done as a precaution incase someone adds a new
job to the workflow but forgets to add needed read permissions.

---------

Co-authored-by: Brandon Diaz <“BrandonLouisDiaz@gmail.com”>