SharedTree: Export base fuzz model (#24933)
The primary goal of this change is to get a base model for tree
exported, so it can be consumed in the local server stress tests. There
are additional issues consuming the tree model in the local server
stress, which i'll address is subsequent PRs. This change matches the
pattern used in our other DDS.

I did make a small change the generator to generate 100 ops, to match
the what is generally used across our stress runs, and this seems to
have uncovered an assert.