fix(handles): Prepare for removal of IFluidHandleInternal.bind (#24863)
This function was deprecated in #24553 and its removal is tracked by
#24557.

This change addresses these last few usages:
* Updated some weird old tests that were calling bind on a DataStore's
handle
* `FluidObjectHandle.bind` implementation, which included tracking
pending handles and such. Copied this up to `SharedObjectHandle` where
it belongs. All that logic will be removed from `FluidObjectHandle` when
`bind` is removed.
* In some DDS tests we test or do serialization, with no interest in
handle binding. But the FluidSerializer implementation requires a valid
bind source. So add on a dummy `bind` function. (right now this is a
no-op since all handles have a `bind` function, but once those are
removed this still needs to work)