fix(container-runtime): Reentrant op rebasing wrongly skipped for single-op batch (#24921)
## Description 

Fixes
[AB#33427](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/33427)

In Outbox.flushInternal, we skip rebase for single-message batches. This
is an eventual consistency bug (Repro here:
https://github.com/microsoft/FluidFramework/compare/main...anthony-murphy:mt-rententrancy).
This change attempts to fix this by removing the message size check when
rebasing in outbox.

The cause of the migration tests that were failing was that In PactMap
(which is used for MigrationTool), the "accept" op is always a single
message reentrant op in response to a "set" op (which proposes a new
version). This "accept" op is now rebased/resubmitted. We had a check in
PactMap that would drop all "accept" ops that were resubmitted because
we assumed that resubmit would only be called in the case of disconnect
(we would drop these because on disconnect our client is removed from
quorum and implicitly accepts). Now accepting the rebasing of single
message batches, we need to adjust this to allow the resubmission of
these "accept" ops.

The solution/change made in PactMap in this PR here is to allow
resubmission of accept ops and passively return early if we receive an
accept op that has already been accepted.