Tree: Local Rollback (#24954)
This change implements local rollback in shared tree and adds some basic
unit test for the scenario. This is modeled after the existing code for
aborting a transaction, which behaves similarly.

This is a precursor change to expanded fuzz coverage for rollback for
all dds via the dds fuzz and incorporating tree into local server
stress. There is a prototype of that here:
https://github.com/microsoft/FluidFramework/pull/24930