ci: Don't remove build-tools build on tag checkout (#24901)
When checking out the newly released tag in the workflow, the checkout
was deleting the previously built build-tools output. Setting `clean:
false` should resolve this, but it doesn't seem to. I switched the steps
to use a `git checkout` command directly, which is both simpler and
works.

I can't explain why this just started happening, expecially since we pin
our action versions. But the pipeline worked in the 2.42.0 release.

---------

Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>