Local Server Stress: Add Tree (#24966)
This pull request enhances the local server stress tests by adding
support for the `tree` data structure and refining the handling of test
options. The most important changes include updating dependencies, and
modifying the test models to include `tree`.

* Imported `baseTreeModel` and added it to the `ddsModelMap`, enabling
the inclusion of `tree` in the test model.
* Refined the logic for handling `only`, `skip`, and `replay` options to
ensure proper precedence and compatibility when seeds overlap.