Update prefetchsnapshot implementation of downloadSnapshot to add CLP Compliant App headers as appropriate (#24987)
## Description

Preteching a snapshot from a MIP protected file with encryption is
failing with insufficient permissions.
Investigation shows that the request is missing the CLP Compliant App
header, and that once one is added the prefetch flow can successfully
download a snapshot for caching.

EpochTracker handles adding the CLP Compatible App Header in the form
body, but some flows, like prefetch do not set EpochTracker. This change
makes so even when epochTracker is not set, the request will contain the
appropriate CLP Compatible app header.

To Repro the issue, just open a Fluid page that is MIP protected and
Encrypted, the overall flow does succeeds, since the fetching of the
snapshot does add an EpochTracker and it handles the CLP header.
But looking at the network trace, one can see a failed fetch triggered
by the prefetchSnapshot call that is missing the CLP header.

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).

Co-authored-by: Jatin Garg <48029724+jatgarg@users.noreply.github.com>