refactor(aqueduct): Simplify `TreeDataObject` and add `TreeDataObjectFactory` (#25013)
Updates `TreeDataObject` to remove the `generateView` pattern, which was
overly prescriptive. `TreeDataObject` no longer supports being directly
SharedTree schema aware. Instead, the implementor is responsible for
applying the right view schema and handling the results accordingly.

In the future, if we decide a general-purpose schema-aware counterpart
type would be useful, we can always add one (without changing the
persisted data format). For now, it is unnecessary, and promotes
obfuscating view schema responsibilities of the implementor.

Also adds `TreeDataObjectFactory` for producing them and updates our
existing `TreeDataObject` sample apps to use the new factory.