Change to createTestLogger (#24880)
## Description

There is a problem in our telemetry where not all of the test results
are being reflected in out portal.
This PR consists in changing the getTestLogger function to the
createTestLogger function which is a necessary part for the solution.

Also updates the pipelines to use `PKG_SPECIFIER` instead of `PKG_PATH`
since it's better to have it consistent throughout our pipelines.

## Reviewer Guidance
Please let me know if there is something else that I should be aware of.

Fixes:
[AB#42164](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/42164)