(compat): Replace Driver layer dependency with Loader layer in the Runtime (#25057)
Runtime communicates with the Driver layer via `IDocumentStorageService`
interface which requires layer compat support across the Runtime /
Driver boundary. However, the Loader layer has a
`ContainerStorageAdapter` which re-implements `IDocumentStorageService`
which is then passed to the Runtime layer.

This PR adds an `IContainerStorageService` interface which is basically
a duplicate of `IDocumentStorageService` and only has properties used by
the `Runtime` layer. `ContainerStorageAdapter` now implements
`IContainerStorageService` and will ensure that it maintains
compatibility with the Runtime layer. So, if anything changes in the
`IDocumentStorageService` interface APIs, `ContainerStorageAdapter` will
still maintain the compatibility with Runtime via
`IContainerStorageService`.

This will remove the Runtime layer's dependency on Driver layer and
replace it with the Loader layer dependency which is already maintained.

Also, added an `IRuntimeStorageService` interface similar to
`IContainerStorageService` to be used by the `DataStore` layer. It will
only have `readBlob` API as the other APIs are not used by `DataStore`
layer.


[AB#33773](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/33773)