refactor(api-markdown-documenter): Align `FencedCodeBlockNode` with Markdown (#25077)
We are working to align this library with Markdown / `mdast`. Fenced
code nodes in `mdast` are literal nodes that wrap a string
representation of the code content. This PR updates
`FencedCodeBlockNode` to use the same pattern and updates the code to
accommodate.

This PR also fixes bugs in both Markdown and HTML rendering of fenced
code blocks where the Markdown renderer was _omitting_ line breaks that
should have been included, and the HTML renderer was (transitively)
causing extra line breaks to be injected.