Fix behavior of object visibility tests with handles in a detached state  (#24997)
#21213 attempted to remove all calls to bindHandles at the DDS layer,
bout could not remove this behavior in directory due to the
fluidObjectVisibility tests. It seemed like there was a bug where a
handle to a detached DDS stored in an intermediate detached DDS would
not be visible from a third detached DDS until the root container
attached.

However, the assumption that a handle would need to be visible (and
therefore able to be resolved) from a root detached container is
incorrect - when detached, each handle has direct access to the object
it points to. Update the visibility tests accordingly, and remove the
call to bindHandles as well.

[AB#8168](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8168)